generalize the magic_key code so that the GTK3 client can re-use it
git-svn-id: https://xpra.org/svn/Xpra/trunk@17149 3bb7dfac-3a0b-4e04-842a-767bc560f471

+++ src/xpra/client/window_border.py
@@ -21,7 +21,7 @@ class WindowBorder(object):
 
     def __repr__(self):
         def hex2(v):
-            b = max(0, min(255, v*256))
+            b = int(max(0, min(255, v*256)))
             if b<16:
                 return "0%X" % b
             return "%X" % b
